Коммит №2 Данные перенесены из фикстуры в файл test_data и исправлена ошибка названия data
