+++ tests/test_changing_user_data.py
@@ -1,13 +1,15 @@
 import requests
 import allure
-from date.urls import changing_data
-from date.generate_unique_user import generate_unique_data
+from data.urls import changing_data
+from data.generate_unique_user import generate_unique_data
+import data.test_data
 
 
 class TestChangingUserData:
 
     @allure.title('Изменение данных name пользователя с авторизацией')
-    def test_with_authorization_change_name(self, auth_token_and_random_user_data):
+    def test_with_authorization_change_name(self):
+        auth_token_and_random_user_data = data.test_data.auth_token_and_random_user_data()
         headers = auth_token_and_random_user_data[0]
         user_data = auth_token_and_random_user_data[1]
         new_user_name = generate_unique_data()
@@ -21,7 +23,8 @@ class TestChangingUserData:
         assert f'{new_name}' == response.json()["user"]["name"] and 200 == response.status_code
 
     @allure.title('Изменение данных email пользователя с авторизацией')
-    def test_with_authorization_change_email(self, auth_token_and_random_user_data):
+    def test_with_authorization_change_email(self):
+        auth_token_and_random_user_data = data.test_data.auth_token_and_random_user_data()
         headers = auth_token_and_random_user_data[0]
         user_data = auth_token_and_random_user_data[1]
         new_user_email = generate_unique_data()
@@ -35,6 +38,7 @@ class TestChangingUserData:
         assert f'{new_email}' == response.json()["user"]["email"] and 200 == response.status_code
 
     @allure.title('Изменение данных пользователя без авторизации')
-    def test_without_authorization(self, payload_registered_user):
+    def test_without_authorization(self):
+        payload_registered_user = data.test_data.payload_registered_user()
         response = requests.patch(changing_data, data=payload_registered_user)
         assert "You should be authorised" == response.json()["message"] and 401 == response.status_code
