+++ tests/test_receiving_user_orders.py
@@ -1,12 +1,14 @@
 import requests
 import allure
-from date.urls import get_order
+import data.test_data
+from data.urls import get_order
 
 
 class TestReceivingUserOrders:
 
     @allure.title('Получение заказов авторизованного пользователя')
-    def test_authorized_user(self, headers_registered_user):
+    def test_authorized_user(self):
+        headers_registered_user = data.test_data.headers_registered_user()
         response = requests.get(get_order, headers=headers_registered_user)
         assert response.json()["success"] and 200 == response.status_code
