+++ tests/test_user_login.py
@@ -1,13 +1,15 @@
 import requests
 import allure
-from date.urls import login_user
-from date.variables import incorrect_login_and_password
+import data.test_data
+from data.urls import login_user
+from data.variables import incorrect_login_and_password
 
 
 class TestUserLogin:
 
     @allure.title('Логин под существующим пользователем')
-    def test_login_existing_user(self, payload_registered_user):
+    def test_login_existing_user(self):
+        payload_registered_user = data.test_data.payload_registered_user()
         response = requests.post(login_user, data=payload_registered_user)
         assert 200 == response.status_code and response.json()["success"]
